+++ src/main/java/bssm/bsmauth/domain/oauth/repositories/OauthTokenRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 @Repository
 public interface OauthTokenRepository extends JpaRepository<OauthToken, String> {
 
-    Optional<OauthToken> findByUsercodeAndClientId(int usercode, String clientId);
+    Optional<OauthToken> findByUsercodeAndClientId(Long usercode, String clientId);
 
     Optional<OauthToken> findByTokenAndExpire(String token, boolean expire);
 }
